+++ test/newsTest.js
@@ -2,22 +2,20 @@ var assert = require("assert");
 var config = require('config');
 var app = require('../app.js');
 var httpRequest = require('supertest')(app);
-var db = require('./DatabaseHelper');
+var db = require('./testRunnerHelper');
 
-var basePath = '/api/v1/tracks';
+var basePath = '/api/v1/news';
 
-describe('-------- TRACK ENDPOINTS --------', function() {
-    var artist;
-    var track;
+describe('-------- NEWS ENDPOINTS --------', function() {
+    var news;
     var token;
 
     //Do preliminary setup, before running each testcase.
     beforeEach(function(done) {
         db.dropAllCollections();
 
-        var options = db.createDateForEndpoint('track');
-        artist =  options.artist;
-        track = options.track;
+        var options = db.createDateForEndpoint('news');
+        news =  options.news;
 
         var tokenPrefix = '?token=';
         token = tokenPrefix + db.createToken().value;
@@ -33,7 +31,7 @@ describe('-------- TRACK ENDPOINTS --------', function() {
     //Testcases
 
     describe('GET ' + basePath, function() {
-        it('Should return a Track item', function(done) {
+        it('Should return a News item', function(done) {
             httpRequest
                 .get(basePath + token)
                 .expect(200)
@@ -43,18 +41,17 @@ describe('-------- TRACK ENDPOINTS --------', function() {
                     if (err)
                         throw err;
 
-                    assert.equal(res.body.result[0]._id, track._id);
-                    assert.equal(res.body.result[0].artists[0].artistName, artist.artistName);
+                    assert.equal(res.body.result[0]._id, news._id);
 
                     done();
                 });
         });
     });
 
-    describe('GET ' + basePath + '/:trackId', function() {
-        it('Should return a Track item whose id is provided', function(done) {
+    describe('GET ' + basePath + '/:newsId', function() {
+        it('Should a News item whose id is provided', function(done) {
             httpRequest
-                .get(basePath + '/' + track._id + token)
+                .get(basePath + '/' + news._id + token)
                 .expect(200)
                 .set('Accept','application/json')
                 .expect('Content-Type', /json/)
@@ -62,18 +59,17 @@ describe('-------- TRACK ENDPOINTS --------', function() {
                     if (err)
                         throw err;
 
-                    assert.equal(res.body.result[0]._id, track._id);
-                    assert.equal(res.body.result[0].artists[0].artistName, artist.artistName);
+                    assert.equal(res.body.result[0]._id, news._id);
 
                     done();
                 });
         });
     });
 
-    describe('GET ' + basePath + '/:artistId/artist', function() {
-        it('Should return a Track item whose id is provided', function(done) {
+    describe('GET ' + basePath + '/:date(dd-mm-yyyy)/date', function() {
+        it('Should a News item whose id is provided', function(done) {
             httpRequest
-                .get(basePath + '/' + artist._id + '/artist' + token)
+                .get(basePath + '/' + news.date + '/date' + token)
                 .expect(200)
                 .set('Accept','application/json')
                 .expect('Content-Type', /json/)
@@ -81,11 +77,13 @@ describe('-------- TRACK ENDPOINTS --------', function() {
                     if (err)
                         throw err;
 
-                    assert.equal(res.body.result[0]._id, track._id);
-                    assert.equal(res.body.result[0].artists[0].artistName, artist.artistName);
+                    assert.equal(res.body.result[0]._id, news._id);
+                    assert.equal(res.body.result[0].date, news.date);
 
                     done();
                 });
         });
     });
+
+
 });
