+++ test/trackTest.js
@@ -2,20 +2,22 @@ var assert = require("assert");
 var config = require('config');
 var app = require('../app.js');
 var httpRequest = require('supertest')(app);
-var db = require('./DatabaseHelper');
+var db = require('./testRunnerHelper');
 
-var basePath = '/api/v1/news';
+var basePath = '/api/v1/tracks';
 
-describe('-------- NEWS ENDPOINTS --------', function() {
-    var news;
+describe('-------- TRACK ENDPOINTS --------', function() {
+    var artist;
+    var track;
     var token;
 
     //Do preliminary setup, before running each testcase.
     beforeEach(function(done) {
         db.dropAllCollections();
 
-        var options = db.createDateForEndpoint('news');
-        news =  options.news;
+        var options = db.createDateForEndpoint('track');
+        artist =  options.artist;
+        track = options.track;
 
         var tokenPrefix = '?token=';
         token = tokenPrefix + db.createToken().value;
@@ -31,7 +33,7 @@ describe('-------- NEWS ENDPOINTS --------', function() {
     //Testcases
 
     describe('GET ' + basePath, function() {
-        it('Should return a News item', function(done) {
+        it('Should return a Track item', function(done) {
             httpRequest
                 .get(basePath + token)
                 .expect(200)
@@ -41,17 +43,18 @@ describe('-------- NEWS ENDPOINTS --------', function() {
                     if (err)
                         throw err;
 
-                    assert.equal(res.body.result[0]._id, news._id);
+                    assert.equal(res.body.result[0]._id, track._id);
+                    assert.equal(res.body.result[0].artists[0].artistName, artist.artistName);
 
                     done();
                 });
         });
     });
 
-    describe('GET ' + basePath + '/:newsId', function() {
-        it('Should a News item whose id is provided', function(done) {
+    describe('GET ' + basePath + '/:trackId', function() {
+        it('Should return a Track item whose id is provided', function(done) {
             httpRequest
-                .get(basePath + '/' + news._id + token)
+                .get(basePath + '/' + track._id + token)
                 .expect(200)
                 .set('Accept','application/json')
                 .expect('Content-Type', /json/)
@@ -59,17 +62,18 @@ describe('-------- NEWS ENDPOINTS --------', function() {
                     if (err)
                         throw err;
 
-                    assert.equal(res.body.result[0]._id, news._id);
+                    assert.equal(res.body.result[0]._id, track._id);
+                    assert.equal(res.body.result[0].artists[0].artistName, artist.artistName);
 
                     done();
                 });
         });
     });
 
-    describe('GET ' + basePath + '/:date(dd-mm-yyyy)/date', function() {
-        it('Should a News item whose id is provided', function(done) {
+    describe('GET ' + basePath + '/:artistId/artist', function() {
+        it('Should return a Track item whose id is provided', function(done) {
             httpRequest
-                .get(basePath + '/' + news.date + '/date' + token)
+                .get(basePath + '/' + artist._id + '/artist' + token)
                 .expect(200)
                 .set('Accept','application/json')
                 .expect('Content-Type', /json/)
@@ -77,13 +81,11 @@ describe('-------- NEWS ENDPOINTS --------', function() {
                     if (err)
                         throw err;
 
-                    assert.equal(res.body.result[0]._id, news._id);
-                    assert.equal(res.body.result[0].date, news.date);
+                    assert.equal(res.body.result[0]._id, track._id);
+                    assert.equal(res.body.result[0].artists[0].artistName, artist.artistName);
 
                     done();
                 });
         });
     });
-
-
 });
